ESQL: Merge function registries
This merges the `FunctionRegistry` from `esql-core` into the
`EsqlFunctionRegistry` in `esql` proper in an effort to shave down the
complexity we got by attempting to share lots of infrastructure with SQL
and KQL. Since we no longer share we can compress these two together.
