+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/plan/logical/meta/MetaFunctions.java
@@ -11,7 +11,6 @@ import org.apache.lucene.util.BytesRef;
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.xpack.esql.core.expression.Attribute;
 import org.elasticsearch.xpack.esql.core.expression.ReferenceAttribute;
-import org.elasticsearch.xpack.esql.core.expression.function.FunctionRegistry;
 import org.elasticsearch.xpack.esql.core.plan.logical.LeafPlan;
 import org.elasticsearch.xpack.esql.core.plan.logical.LogicalPlan;
 import org.elasticsearch.xpack.esql.core.tree.NodeInfo;
@@ -49,7 +48,7 @@ public class MetaFunctions extends LeafPlan {
         return attributes;
     }
 
-    public List<List<Object>> values(FunctionRegistry functionRegistry) {
+    public List<List<Object>> values(EsqlFunctionRegistry functionRegistry) {
         List<List<Object>> rows = new ArrayList<>();
         for (var def : functionRegistry.listFunctions(null)) {
             EsqlFunctionRegistry.FunctionDescription signature = EsqlFunctionRegistry.description(def);
