+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/expression/function/EsqlFunctionRegistryTests.java
@@ -12,13 +12,12 @@ import org.elasticsearch.compute.operator.EvalOperator;
 import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.xpack.esql.core.ParsingException;
 import org.elasticsearch.xpack.esql.core.QlIllegalArgumentException;
+import org.elasticsearch.xpack.esql.core.TestUtils;
 import org.elasticsearch.xpack.esql.core.expression.Expression;
 import org.elasticsearch.xpack.esql.core.expression.function.FunctionDefinition;
-import org.elasticsearch.xpack.esql.core.expression.function.FunctionRegistry;
-import org.elasticsearch.xpack.esql.core.expression.function.FunctionRegistryTests;
 import org.elasticsearch.xpack.esql.core.expression.function.FunctionResolutionStrategy;
-import org.elasticsearch.xpack.esql.core.expression.function.OptionalArgument;
 import org.elasticsearch.xpack.esql.core.expression.function.UnresolvedFunction;
+import org.elasticsearch.xpack.esql.core.expression.function.scalar.ScalarFunction;
 import org.elasticsearch.xpack.esql.core.session.Configuration;
 import org.elasticsearch.xpack.esql.core.tree.NodeInfo;
 import org.elasticsearch.xpack.esql.core.tree.Source;
@@ -26,22 +25,92 @@ import org.elasticsearch.xpack.esql.core.tree.SourceTests;
 import org.elasticsearch.xpack.esql.core.type.DataType;
 import org.elasticsearch.xpack.esql.expression.function.scalar.EsqlConfigurationFunction;
 
+import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 import java.util.function.Function;
 
+import static java.util.Collections.emptyList;
 import static org.elasticsearch.xpack.esql.EsqlTestUtils.randomConfiguration;
-import static org.elasticsearch.xpack.esql.core.expression.function.FunctionRegistry.def;
 import static org.elasticsearch.xpack.esql.core.expression.function.FunctionResolutionStrategy.DEFAULT;
+import static org.elasticsearch.xpack.esql.expression.function.EsqlFunctionRegistry.def;
 import static org.hamcrest.Matchers.endsWith;
 import static org.hamcrest.Matchers.is;
 import static org.mockito.Mockito.mock;
 
 public class EsqlFunctionRegistryTests extends ESTestCase {
 
+    public void testNoArgFunction() {
+        UnresolvedFunction ur = uf(DEFAULT);
+        EsqlFunctionRegistry r = new EsqlFunctionRegistry(def(DummyFunction.class, DummyFunction::new, "dummyFunction"));
+        FunctionDefinition def = r.resolveFunction(ur.name());
+        assertEquals(ur.source(), ur.buildResolved(TestUtils.randomConfiguration(), def).source());
+    }
+
+    public void testBinaryFunction() {
+        UnresolvedFunction ur = uf(DEFAULT, mock(Expression.class), mock(Expression.class));
+        EsqlFunctionRegistry r = new EsqlFunctionRegistry(def(DummyFunction.class, (Source l, Expression lhs, Expression rhs) -> {
+            assertSame(lhs, ur.children().get(0));
+            assertSame(rhs, ur.children().get(1));
+            return new DummyFunction(l);
+        }, "dummyFunction"));
+        FunctionDefinition def = r.resolveFunction(ur.name());
+        assertEquals(ur.source(), ur.buildResolved(TestUtils.randomConfiguration(), def).source());
+
+        // No children aren't supported
+        ParsingException e = expectThrows(ParsingException.class, () -> uf(DEFAULT).buildResolved(TestUtils.randomConfiguration(), def));
+        assertThat(e.getMessage(), endsWith("expects exactly two arguments"));
+
+        // One child isn't supported
+        e = expectThrows(
+            ParsingException.class,
+            () -> uf(DEFAULT, mock(Expression.class)).buildResolved(TestUtils.randomConfiguration(), def)
+        );
+        assertThat(e.getMessage(), endsWith("expects exactly two arguments"));
+
+        // Many children aren't supported
+        e = expectThrows(
+            ParsingException.class,
+            () -> uf(DEFAULT, mock(Expression.class), mock(Expression.class), mock(Expression.class)).buildResolved(
+                TestUtils.randomConfiguration(),
+                def
+            )
+        );
+        assertThat(e.getMessage(), endsWith("expects exactly two arguments"));
+    }
+
+    public void testAliasNameIsTheSameAsAFunctionName() {
+        EsqlFunctionRegistry r = new EsqlFunctionRegistry(def(DummyFunction.class, DummyFunction::new, "DUMMY_FUNCTION", "ALIAS"));
+        QlIllegalArgumentException iae = expectThrows(
+            QlIllegalArgumentException.class,
+            () -> r.register(def(DummyFunction2.class, DummyFunction2::new, "DUMMY_FUNCTION2", "DUMMY_FUNCTION"))
+        );
+        assertEquals("alias [DUMMY_FUNCTION] is used by [DUMMY_FUNCTION] and [DUMMY_FUNCTION2]", iae.getMessage());
+    }
+
+    public void testDuplicateAliasInTwoDifferentFunctionsFromTheSameBatch() {
+        QlIllegalArgumentException iae = expectThrows(
+            QlIllegalArgumentException.class,
+            () -> new EsqlFunctionRegistry(
+                def(DummyFunction.class, DummyFunction::new, "DUMMY_FUNCTION", "ALIAS"),
+                def(DummyFunction2.class, DummyFunction2::new, "DUMMY_FUNCTION2", "ALIAS")
+            )
+        );
+        assertEquals("alias [ALIAS] is used by [DUMMY_FUNCTION(ALIAS)] and [DUMMY_FUNCTION2]", iae.getMessage());
+    }
+
+    public void testDuplicateAliasInTwoDifferentFunctionsFromTwoDifferentBatches() {
+        EsqlFunctionRegistry r = new EsqlFunctionRegistry(def(DummyFunction.class, DummyFunction::new, "DUMMY_FUNCTION", "ALIAS"));
+        QlIllegalArgumentException iae = expectThrows(
+            QlIllegalArgumentException.class,
+            () -> r.register(def(DummyFunction2.class, DummyFunction2::new, "DUMMY_FUNCTION2", "ALIAS"))
+        );
+        assertEquals("alias [ALIAS] is used by [DUMMY_FUNCTION] and [DUMMY_FUNCTION2]", iae.getMessage());
+    }
+
     public void testFunctionResolving() {
         UnresolvedFunction ur = uf(DEFAULT, mock(Expression.class));
-        FunctionRegistry r = new EsqlFunctionRegistry(defineDummyFunction(ur, "dummyfunction", "dummyfunc"));
+        EsqlFunctionRegistry r = new EsqlFunctionRegistry(defineDummyFunction(ur, "dummyfunction", "dummyfunc"));
 
         // Resolve by primary name
         FunctionDefinition def;
@@ -72,7 +141,7 @@ public class EsqlFunctionRegistryTests extends ESTestCase {
 
     public void testUnaryFunction() {
         UnresolvedFunction ur = uf(DEFAULT, mock(Expression.class));
-        FunctionRegistry r = new EsqlFunctionRegistry(defineDummyUnaryFunction(ur));
+        EsqlFunctionRegistry r = new EsqlFunctionRegistry(defineDummyUnaryFunction(ur));
         FunctionDefinition def = r.resolveFunction(ur.name());
 
         // No children aren't supported
@@ -90,8 +159,8 @@ public class EsqlFunctionRegistryTests extends ESTestCase {
     public void testConfigurationOptionalFunction() {
         UnresolvedFunction ur = uf(DEFAULT, mock(Expression.class));
         FunctionDefinition def;
-        FunctionRegistry r = new EsqlFunctionRegistry(
-            EsqlFunctionRegistry.def(DummyConfigurationOptionalArgumentFunction.class, (Source l, Expression e, Configuration c) -> {
+        EsqlFunctionRegistry r = new EsqlFunctionRegistry(
+            def(DummyConfigurationOptionalArgumentFunction.class, (Source l, Expression e, Configuration c) -> {
                 assertSame(e, ur.children().get(0));
                 return new DummyConfigurationOptionalArgumentFunction(l, List.of(ur), c);
             }, "dummy")
@@ -105,9 +174,9 @@ public class EsqlFunctionRegistryTests extends ESTestCase {
     }
 
     private static FunctionDefinition defineDummyFunction(UnresolvedFunction ur, String... names) {
-        return def(FunctionRegistryTests.DummyFunction.class, (Source l, Expression e) -> {
+        return def(DummyFunction.class, (Source l, Expression e) -> {
             assertSame(e, ur.children().get(0));
-            return new FunctionRegistryTests.DummyFunction(l);
+            return new DummyFunction(l);
         }, names);
     }
 
@@ -125,6 +194,43 @@ public class EsqlFunctionRegistryTests extends ESTestCase {
             }
         }
         return output.toString();
+    }
+
+    public static class DummyFunction extends ScalarFunction {
+        public DummyFunction(Source source) {
+            super(source, emptyList());
+        }
+
+        @Override
+        public void writeTo(StreamOutput out) throws IOException {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public String getWriteableName() {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        protected NodeInfo<DummyFunction> info() {
+            return NodeInfo.create(this);
+        }
+
+        @Override
+        public Expression replaceChildren(List<Expression> newChildren) {
+            throw new UnsupportedOperationException("this type of node doesn't have any children to replace");
+        }
+
+        @Override
+        public DataType dataType() {
+            return null;
+        }
+    }
+
+    public static class DummyFunction2 extends DummyFunction {
+        public DummyFunction2(Source source) {
+            super(source);
+        }
     }
 
     public static class DummyConfigurationOptionalArgumentFunction extends EsqlConfigurationFunction implements OptionalArgument {
